requirements.txt
adbutils==2.7.2
alembic==1.13.1
allure-pytest==2.13.5
allure-python-commons==2.13.5
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
apkutils2==1.0.0
Appium-Python-Client==4.0.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cigam==0.0.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configparser==7.0.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
decorator==5.1.1
deprecation==2.1.0
Django==5.0.7
django-extensions==3.2.3
django-redis==5.4.0
django-sslserver==0.22
django-sslserver2==0.1.4
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
Faker==26.1.0
fastapi==0.111.0
fastapi-cli==0.0.2
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
iniconfig==2.0.0
Jinja2==3.1.3
kombu==5.3.7
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.1
pillow==10.3.0
playwright==1.45.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyelftools==0.31
Pygments==2.18.0
pymongo==4.8.0
pyOpenSSL==24.2.1
PySocks==1.7.1
pytest==8.2.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
retry==0.9.2
rich==13.7.1
selenium==4.21.0
service-identity==24.1.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
sqlmodel==0.0.18
sqlparse==0.5.0
starlette==0.37.2
text-unidecode==1.3
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
uiautomator2==3.1.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
xmltodict==0.13.0
zope.interface==6.4.post2
